requirements.txt
hidapi==0.14.0
psutil==5.9.5
